+++ projects/egeo/src/lib/st-dynamic-table/utils/st-dynamic-table.utils.ts
@@ -39,6 +39,7 @@ export class StDynamicTableUtils {
                         filters: filters,
                         filterable: filters && (filters.templateRef !== undefined || (filters.filterConfig && filters.filterConfig.length > 0)),
                         sortable: isSortable,
+                        sortedByDefault: uiDefinition && uiDefinition.sort ? uiDefinition.sort.toUpperCase() : null,
                         type: _property.type ? this._getTypes(_propertyKey, _property.type.toString(), jsonSchema, uiDefinition) : null
                      });
                   }