+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * st-form: added .small class to affect form inner inputs or input components individually.
 * st-dynamic-table: Allow to hide a field using the UI definitions
+* st-dynamic-table: Manage default ordering in table component
 * st-table: Allow to display "sticky" the hover menu
 * st-table: Close filter menu when user clicks outside